move state func to state.go
